+++ src/pages/signup/components/SignupSuccess.jsx
@@ -1,0 +1,10 @@
+const SignupSuccess = () => {
+  return (
+    <>
+      <h1>Cadastro realizado com sucesso!</h1>
+      <p>Por favor, verifique seu e-mail e siga as instruções enviadas para validar sua conta e começar a utilizar a Kanboom</p>
+    </>
+  );
+};
+
+export default SignupSuccess;
